requirements.txt
asgiref==3.7.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-import-export==3.3.1
django-redis==5.3.0
django-rq==2.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
rq==1.15.1
sqlparse==0.4.4
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
